Remove wrapper exe
